// Desktop/BIKE/bicycle/src/Shared/Types/Member.tsx
export interface User {
    _id: string;

  name: string;

  email: string;

  password: string;

  firstname: string;

  lastname: string;

  phoneNumber: string;

  address: string;

  city: string;

  street: string;

  house: string;

  apartment: string;

  detailedInformation: string;

  role: string;

  status: string;
}